Working backend get requests

+++ backend/src/main/java/com/example/madrid/madrid/EstablishmentRepository.java
@@ -2,7 +2,7 @@ package com.example.madrid.madrid;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface EstablishmentRepository extends CrudRepository<Establishment, Integer> {
+public interface EstablishmentRepository extends CrudRepository<Establishment, Long> {
     Establishment findByName(String name);
 }
